Fix cut -> cd somedir -> paste not updating `files`
The problem:
When
    set period 0
    set watch true
    set dircache false

One could copy/cut and paste a file, and `files` wont be updated
until `:reload` is called, despite it being visible in the
directory.

Likely because with `gOpts.watch == true` ":load" isn't executed.

`exportFilesInCurrDir()` should be after `for i := range app.nav.dirs`,
but I put at the very end just in case.